+++ R/PIOutputMapping.R
@@ -8,7 +8,6 @@
 #' @format NULL
 PIOutputMapping <- R6::R6Class(
   "PIOutputMapping",
-  inherit = ospsuite.utils::Printable,
   cloneable = TRUE,
   active = list(
     #' @field observedDataSets A named list containing `DataSet` objects for comparison
@@ -121,7 +120,8 @@ PIOutputMapping <- R6::R6Class(
           result <- try(
             {
               data[[idx]]$molWeight <- ospsuite::getMolWeightFor(
-                private$.quantity, unit = "g/mol"
+                private$.quantity,
+                unit = "g/mol"
               )
               tryConvert()
             },
@@ -197,11 +197,15 @@ PIOutputMapping <- R6::R6Class(
 
     #' @description Prints a summary of the PIOutputMapping.
     print = function() {
-      private$printClass()
-      private$printLine("Output path", private$.quantity$path)
-      private$printLine("Observed data labels", names(private$.observedDataSets))
-      private$printLine("Scaling", private$.scaling)
-      invisible(self)
+      ospsuite.utils::ospPrintClass(self)
+      ospsuite.utils::ospPrintItems(
+        list(
+          "Output path" = private$.quantity$path,
+          "Observed data labels" = names(private$.observedDataSets),
+          "Scaling" = private$.scaling
+        ),
+        print_empty = TRUE
+      )
     }
   )
 )